refactor(T1 E1): extract Scaled Extend into USFExtendScaledService [BUILD-PASSED]
Slice E1 of the Global Sequencing plan. Moves the Scaled Extend planning/preview/validate
methods (OnScaledExtendStateChanged, CalculateScaledExtendPositions, SpawnScaledExtendPreviews,
ClearScaledExtendClones, ValidateScaledExtendConstraints, ValidatePowerCapacity) into a new
USFExtendScaledService. SFExtendService.cpp 7,718 -> 6,436 (-1,282).

Careful move (round-1 friend/Owner-> pattern): all state stays on USFExtendService; the sub-
service is a friend and operates on ScaledExtendClones/StoredCloneTopology/Json* maps/
HologramService in place via Owner->. FSFScaledExtendClone relocated from a nested struct to
SFExtendScaledService.h (also used by the post-build wiring path). Thin forwarders kept for
OnScaledExtendStateChanged/ClearScaledExtendClones/ValidatePowerCapacity (Subsystem + internal
callers unchanged); the 3 tiny getters (GetExtendCloneCount/GetExtendRowCount/IsScaledExtendActive)
kept in place. Shipping PackagePlugin compile+cook clean (ExitCode=0, BUILD SUCCESSFUL).
Static-verified: 0 unqualified members, 0 bare this, string literals protected during qualification.
In-game smoke pending.
